+++ demo/ggforce.R
@@ -2,8 +2,8 @@
 library(ggforce)
 library(units)
 gallon = as_units("gallon")
-mtcars$consumption = mtcars$mpg * with(ud_units, mi/gallon)
-mtcars$power = mtcars$hp * with(ud_units, hp)
+mtcars$consumption = mtcars$mpg * make_units(mi/gallon)
+mtcars$power = mtcars$hp * make_units(hp)
 ggplot(mtcars) + geom_point(aes(power, consumption))
 ggplot(mtcars) + geom_point(aes(1/power, 1/consumption))
 
@@ -13,10 +13,10 @@ ggplot(mtcars) + geom_point(aes(1/power, 1/consumption))
 library(units)
 units_options(negative_power = TRUE)
 gallon = as_units("gallon")
-mtcars$consumption = mtcars$mpg * with(ud_units, mi/gallon)
-mtcars$displacement = mtcars$disp * ud_units[["in"]]^3
-units(mtcars$displacement) = with(ud_units, cm^3)
-units(mtcars$consumption) = with(ud_units, km/l)
+mtcars$consumption = mtcars$mpg * make_units(mi/gallon)
+mtcars$displacement = mtcars$disp * as_units("in")^3
+units(mtcars$displacement) = make_units(cm^3)
+units(mtcars$consumption) = make_units(km/l)
 m = par("mar")
 m[3] = 0.1
 par(mar = m + c(0, .3, 0, 0))