+++ tests/testthat/test_udunits.R
@@ -42,6 +42,4 @@ test_that("udunits database can be read", {
   skip_if_not_installed("xml2")
   expect_message(valid_udunits(), "udunits")
   expect_message(valid_udunits_prefixes(), "udunits")
-  # expect_identical(units:::.construct_ud_units(), ud_units) # NOT TRUE -- *ppv not in ud_units?
-  expect_silent(units:::.construct_ud_units())
 })